Revert "Feature editing"
