+++ package.json
@@ -5,15 +5,11 @@
   "proxy": "http://localhost:5000",
   "dependencies": {
     "@material-ui/core": "^3.9.4",
-    "@material-ui/icons": "^4.9.1",
     "axios": "^0.18.1",
-    "body-parser": "^1.19.0",
-    "express": "^4.17.1",
     "pg": "^7.18.2",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
     "react-redux": "^5.1.2",
-    "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.0",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
