--- src/components/Gallery/Gallery.css
@@ -1,6 +1,0 @@
-.GalleryDiv {
-  display: flex;
-  width: 100%;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-}
