--- src/components/Gallery/GalleryItem/GalleryItem.css
@@ -1,21 +1,0 @@
-.GalleryItemContainer {
-  display: grid;
-  grid:
-    "img    title" auto
-    "img    desc" 1fr
-    "img    desc" 1fr / 185px 1fr;
-}
-
-.GalleryItemImg {
-  grid-area: img;
-}
-
-.GalleryItemTitle {
-  grid-area: title;
-  margin-bottom: 0;
-}
-
-.GalleryItemDescription {
-  grid-area: desc;
-  text-align: left;
-}
