--- src/reducers/movies.js
@@ -1,13 +1,0 @@
-/**
- * Used to store movies returned from the server.
- */
-const movies = (state = [], action) => {
-  switch (action.type) {
-    case "SET_MOVIES":
-      return action.payload;
-    default:
-      return state;
-  }
-};
-
-export default movies;
