--- src/views/ListPage.js
@@ -1,15 +1,0 @@
-import React, { Component } from "react";
-import Gallery from "../components/Gallery/Gallery";
-
-class ListPage extends Component {
-  render() {
-    return (
-      <div>
-        <h2>LIST PAGE</h2>
-        <Gallery />
-      </div>
-    );
-  }
-}
-
-export default ListPage;
